+++ lumin/nn/callbacks/opt_callbacks.py
@@ -1,6 +1,7 @@
 import numpy as np
 import math
 from typing import Tuple, Optional
+import pandas as pd
 
 from .callback import Callback
 from ..models.abs_model import AbsModel
@@ -28,7 +29,7 @@ class LRFinder(Callback):
         self.lr_bounds = lr_bounds
         self.lr_mult = (self.lr_bounds[1]/self.lr_bounds[0])**(1/nb)
         
-    def on_train_begin(self, **kargs):
+    def on_train_begin(self, **kargs) -> None:
         r'''
         Prepares variables and optimiser for new training
         '''
@@ -39,7 +40,7 @@ class LRFinder(Callback):
         
     def _calc_lr(self): return self.lr_bounds[0]*(self.lr_mult**self.iter)
     
-    def plot(self, n_skip:int=0, n_max:Optional[int]=None, lim_y:Optional[Tuple[float,float]]=None):
+    def plot(self, n_skip:int=0, n_max:Optional[int]=None, lim_y:Optional[Tuple[float,float]]=None) -> None:
         r'''
         Plot the loss as a function of the LR.
 
@@ -48,6 +49,8 @@ class LRFinder(Callback):
             n_max: Maximum iteration number to plot
             lim_y: y-range for plotting
         '''
+
+        # TODO: Decide on whether to keep this; could just pass to plot_lr_finders
 
         with sns.axes_style(self.plot_settings.style), sns.color_palette(self.plot_settings.cat_palette):
             plt.figure(figsize=(self.plot_settings.w_mid, self.plot_settings.h_mid))
@@ -62,7 +65,7 @@ class LRFinder(Callback):
             plt.xlabel("Learning rate", fontsize=self.plot_settings.lbl_sz, color=self.plot_settings.lbl_col)
             plt.show()
         
-    def plot_lr(self):
+    def plot_lr(self) -> None:
         r'''
         Plot the LR as a function of iterations.
         '''
@@ -76,7 +79,14 @@ class LRFinder(Callback):
             plt.xlabel("Iterations", fontsize=self.plot_settings.lbl_sz, color=self.plot_settings.lbl_col)
             plt.show()
 
-    def on_batch_end(self, loss:float, **kargs):
+    def get_df(self) -> pd.DataFrame:
+        r'''
+        Returns a DataFrame of LRs and losses
+        '''
+
+        return pd.DataFrame({'LR': self.history['lr'], 'Loss': self.history['loss']})
+
+    def on_batch_end(self, loss:float, **kargs) -> None:
         r'''
         Records loss and increments LR