+++ CHANGES.md
@@ -7,6 +7,12 @@
 ## Additions
 
 - `rf_check_feat_removal` method to check whether one of several (correlated) features can safely be ignored
+- `rf_rank_features`:
+
+    -`n_max_display` to `rf_rank_features` to adjucst number of features displayed in plot
+    - `plot_results`, `retrain_on_import_feats`, and `verbose` to control printed outputs of function
+- Control over x-axis label in `plot_importance`
+- `repeated_rf_rank_features`
 
 ## Removals
 
@@ -15,6 +21,7 @@
 - Remove mutable default paramert for `get_opt_rf_params`
 - Missing `n_estimators` in call to `get_opt_rf_params` to `rf_rank_features`
 - Added string interpretation check when loading `ModelBuilder` saved in pre-v0.3.1 versions
+- `rf_rank_features` importance cut now >= threshold, was previously >
 
 ## Changes